Moved and renamed files, cleaned up notebook
