+++ tasks/fine_tuning_sbert/src/sentence_transformer.py
@@ -5,50 +5,33 @@ Implementing the Early Stopping feature that will be useful for us
 Original source code: https://github.com/UKPLab/sentence-transformers/blob/master/sentence_transformers/SentenceTransformer.py#L434
 """
 
-import math, time
 from typing import Iterable, Dict, Tuple, Type, Callable
-import pandas as pd
-import sys, os, csv, json, random, queue
-
-from sklearn.model_selection import train_test_split
-from lightgbm import LGBMClassifier
-from sklearn import svm
-from sklearn.metrics import classification_report
+import os
 import transformers
 import matplotlib.pyplot as plt
-from sklearn.ensemble import RandomForestClassifier
-from sentence_transformers import SentencesDataset, InputExample, losses, SentenceTransformer
-from sentence_transformers.evaluation import LabelAccuracyEvaluator
-from torch import nn, Tensor
-from torch.utils.data import DataLoader
-import cupy as cp
-import cupy as cp
-import spacy
+from sentence_transformers import SentenceTransformer
 
 from sentence_transformers.evaluation import LabelAccuracyEvaluator, SentenceEvaluator
-from sklearn.metrics import classification_report
-from sklearn.model_selection import train_test_split
-from torch import nn, Tensor
+from torch import nn
 import torch
 from torch.optim import Optimizer
 from torch.utils.data import DataLoader
-from torch import device
-import torch.multiprocessing as mp
 from tqdm.autonotebook import trange
 from statistics import mean
 
 
-class EarlyStoppingSentenceTransformer(SentenceTransformer): # I checked and I think the same name will work when overriding the fit function
+# I checked and I think the same name will work when overriding the fit function
+class EarlyStoppingSentenceTransformer(SentenceTransformer):
 
     def fit(self,
             train_objectives: Iterable[Tuple[DataLoader, nn.Module]],
             evaluator: SentenceEvaluator = None,
             epochs: int = 1,
-            steps_per_epoch = None,
+            steps_per_epoch=None,
             scheduler: str = 'WarmupLinear',
             warmup_steps: int = 10000,
             optimizer_class: Type[Optimizer] = transformers.AdamW,
-            optimizer_params : Dict[str, object]= {'lr': 2e-5},
+            optimizer_params: Dict[str, object] = {'lr': 2e-5},
             weight_decay: float = 0.01,
             evaluation_steps: int = 0,
             output_path: str = None,
@@ -88,10 +71,11 @@ class EarlyStoppingSentenceTransformer(SentenceTransformer): # I checked and I t
         :param PATIENCE: maximum number of epochs to go without an improvement in the accuracy
         :param params: contains the name of the model, test percentage, and seed to save the accuracy plot in drive
         """
-        self.ACC_LIST = [1e-6] # stores the accuracy while training
+        self.ACC_LIST = [1e-6]  # stores the accuracy while training
         TRAINING_ACC_LIST = []
 
-        t_evaluator = LabelAccuracyEvaluator(dataloader=train_objectives[0][0], softmax_model=train_objectives[0][1], name='lae-training')
+        t_evaluator = LabelAccuracyEvaluator(dataloader=train_objectives[0][0], softmax_model=train_objectives[0][1],
+                                             name='lae-training')
 
         self.BASELINE = BASELINE
         self.PATIENCE = PATIENCE
@@ -130,16 +114,17 @@ class EarlyStoppingSentenceTransformer(SentenceTransformer): # I checked and I t
 
             no_decay = ['bias', 'LayerNorm.bias', 'LayerNorm.weight']
             optimizer_grouped_parameters = [
-                {'params': [p for n, p in param_optimizer if not any(nd in n for nd in no_decay)], 'weight_decay': weight_decay},
+                {'params': [p for n, p in param_optimizer if not any(nd in n for nd in no_decay)],
+                 'weight_decay': weight_decay},
                 {'params': [p for n, p in param_optimizer if any(nd in n for nd in no_decay)], 'weight_decay': 0.0}
             ]
 
             optimizer = optimizer_class(optimizer_grouped_parameters, **optimizer_params)
-            scheduler_obj = self._get_scheduler(optimizer, scheduler=scheduler, warmup_steps=warmup_steps, t_total=num_train_steps)
+            scheduler_obj = self._get_scheduler(optimizer, scheduler=scheduler, warmup_steps=warmup_steps,
+                                                t_total=num_train_steps)
 
             optimizers.append(optimizer)
             schedulers.append(scheduler_obj)
-
 
         global_step = 0
         data_iterators = [iter(dataloader) for dataloader in dataloaders]
@@ -168,9 +153,7 @@ class EarlyStoppingSentenceTransformer(SentenceTransformer): # I checked and I t
                         data_iterators[train_idx] = data_iterator
                         data = next(data_iterator)
 
-
                     features, labels = data
-
 
                     if use_amp:
                         with autocast():
@@ -208,34 +191,36 @@ class EarlyStoppingSentenceTransformer(SentenceTransformer): # I checked and I t
 
             # validation evaluation
             flag = self._eval_during_training(evaluator, output_path, save_best_model, epoch, -1, callback)
-            
+
             if flag is True:
                 print(f'Epoch: {epoch}')
                 print(f"Best score: {self.best_score}")
-                print('='*60)
+                print('=' * 60)
             else:
                 print('TRAINING EXITED. Best model has been found.')
                 print(f'Epoch: {epoch}')
                 print(f"Best score: {self.best_score}")
-                print('='*60)
-                plt.plot(range(1, len(TRAINING_ACC_LIST)+1), TRAINING_ACC_LIST, 'orange', label='Training accuracy')
-                plt.plot(range(1, len(self.ACC_LIST)+1), self.ACC_LIST, 'blue', label='Validation accuracy')
+                print('=' * 60)
+                plt.plot(range(1, len(TRAINING_ACC_LIST) + 1), TRAINING_ACC_LIST, 'orange', label='Training accuracy')
+                plt.plot(range(1, len(self.ACC_LIST) + 1), self.ACC_LIST, 'blue', label='Validation accuracy')
                 plt.title('Training and Validation accuracy')
                 plt.xlabel('Epochs')
                 plt.ylabel('Accuracy')
                 plt.rcParams["figure.figsize"] = [15, 15]
                 plt.legend()
-                plt.savefig(f"{output_path}/SBERT_model={params['model_name']}_test_perc={params['test_perc']}_n_epochs={epochs}_seed={params['seed']}.png", bbox_inches ="tight",)
-                plt.show()                
+                plt.savefig(
+                    f"{output_path}/SBERT_model={params['model_name']}_test_perc={params['test_perc']}_n_epochs={epochs}_seed={params['seed']}.png",
+                    bbox_inches="tight", )
+                plt.show()
                 return
-            
+
             # removing the unnecessary first element in ACC_LIST that needed to be there for epoch 1
             if epoch == 0:
                 del self.ACC_LIST[0]
-                
-        if evaluator is None and output_path is not None:   #No evaluator, but output path: save final model version
+
+        if evaluator is None and output_path is not None:  # No evaluator, but output path: save final model version
             self.save(output_path)
-    
+
     def _eval_during_training(self, evaluator, output_path, save_best_model, epoch, steps, callback):
         """Runs evaluation during the training"""
 
@@ -244,24 +229,25 @@ class EarlyStoppingSentenceTransformer(SentenceTransformer): # I checked and I t
         self.ACC_LIST.append(score)
 
         prev_score = self.ACC_LIST[-2]
-        moving_average = mean(self.ACC_LIST[-self.PATIENCE-1: -1])
+        moving_average = mean(self.ACC_LIST[-self.PATIENCE - 1: -1])
 
-        print(f"{'='*60}\nCurrent Score is: {score}\nCurrent ACC_LIST is: {self.ACC_LIST}")
-        
-        if score >= moving_average or len(self.ACC_LIST) - 1 <= self.PATIENCE: # score is >= the moving average in the last PATIENCE values
-            if score > prev_score and score - prev_score >= self.BASELINE: # better score 
-                if score > self.best_score: # checking for local maxima
+        print(f"{'=' * 60}\nCurrent Score is: {score}\nCurrent ACC_LIST is: {self.ACC_LIST}")
+
+        if score >= moving_average or len(
+                self.ACC_LIST) - 1 <= self.PATIENCE:  # score is >= the moving average in the last PATIENCE values
+            if score > prev_score and score - prev_score >= self.BASELINE:  # better score
+                if score > self.best_score:  # checking for local maxima
                     self.best_score = score
                     self.save(output_path)
-                return True # continue training whether this is local maxima or not
+                return True  # continue training whether this is local maxima or not
             elif score >= prev_score and score - prev_score < self.BASELINE:
-                if score > self.best_score: # checking for local maxima
+                if score > self.best_score:  # checking for local maxima
                     self.best_score = score
                     self.save(output_path)
-                return False # end training whether this is local maxima or not, no more training happening after this plateau
+                return False  # end training whether this is local maxima or not, no more training happening after this plateau
             else:
                 # if current score < previous score
-                return True # do not save the model but continue training
+                return True  # do not save the model but continue training
         else:
             print(f'Current score ({score}) less than moving average ({moving_average})')
-            return False # if this accuracy is less than moving average, we do not want to save the weights of this epoch
+            return False  # if this accuracy is less than moving average, we do not want to save the weights of this epoch
